feat: 데이터 패칭 함수 분리, 에러 핸들링 적용
