+++ src/app/_hooks/useFetchData.ts
@@ -1,4 +1,5 @@
 import { ERROR_MESSAGE, GC_TIME, STALE_TIME } from '@/config/const';
+import { filteredDataApi } from '@/service/service';
 import { OrgCourseListResponse } from '@/types/const';
 import { useQuery } from '@tanstack/react-query';
 import { v4 as uuidv4 } from 'uuid';
@@ -13,27 +14,28 @@ export const useFetchData = (
   const { data } = useQuery<OrgCourseListResponse>({
     queryKey: ['selected', searchWord, mappedKey, offset, count],
     queryFn: async () => {
-      const response = await fetch(
-        `http://localhost:3000/api/get/list?encodeUrl=${encodeUrl}&offset=${offset}&count=${count}`,
-        {
-          method: 'GET',
-          cache: 'no-store',
-        },
-      );
-      if (response.ok) {
-        const jsonData: OrgCourseListResponse = await response.json();
-        const coursesWithId = jsonData.courses.map((course) => ({
-          ...course,
-          id: uuidv4(),
-        }));
+      return await filteredDataApi.getFilteredData(encodeUrl, offset, count);
+      // const response = await fetch(
+      //   `http://localhost:3000/api/get/list?encodeUrl=${encodeUrl}&offset=${offset}&count=${count}`,
+      //   {
+      //     method: 'GET',
+      //     cache: 'no-store',
+      //   },
+      // );
+      // if (response.ok) {
+      //   const jsonData: OrgCourseListResponse = await response.json();
+      //   const coursesWithId = jsonData.courses.map((course) => ({
+      //     ...course,
+      //     id: uuidv4(),
+      //   }));
 
-        return {
-          ...jsonData,
-          courses: coursesWithId,
-        };
-      }
-      const errorText = await response.text();
-      throw new Error(errorText || ERROR_MESSAGE.OHTER);
+      //   return {
+      //     ...jsonData,
+      //     courses: coursesWithId,
+      //   };
+      // }
+      // const errorText = await response.text();
+      // throw new Error(errorText || ERROR_MESSAGE.OHTER);
     },
     staleTime: STALE_TIME,
     gcTime: GC_TIME,
